Add A+ utils file
